requirements.txt
Flask==3.0.3
Pillow==11.0.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
